+++ Runtime/EventManager.cs
@@ -30,7 +30,7 @@ namespace ChickadeeEvents
         public List<Rule> rules = new List<Rule>();
 
         public float debugEventDelay;
-        public List<string> debugLog = new List<string>();
+        public List<EventCall> debugLog = new List<EventCall>();
 
         private void Start()
         {
@@ -38,15 +38,13 @@ namespace ChickadeeEvents
                 _current = this;
             else if (_current != this)
                 Destroy(gameObject);
-
-            CallEvent(new EventCall("on_say_dialogue", "player", "mayor"));
         }
 
         public void CallEvent(EventCall call)
         {
-            EventQuery query = new EventQuery(call.eventName, blackboard, call.eventFacts);
+            EventQuery query = new EventQuery(call.EventName, blackboard, call.eventFacts);
 
-            debugLog.Add($"[{Time.time}] {call}");
+            debugLog.Add(call);
 
             OnCallEvent?.Invoke(query);
 
@@ -63,8 +61,8 @@ namespace ChickadeeEvents
                             derefedFacts.Add(new Fact(f.key, query.Deref(f.value)));
                         }
                         EventCall derefedCall = new EventCall(
-                                        query.Deref(response.eventName),
-                                        derefedFacts);
+                                        query.Deref(response.EventName),
+                                        derefedFacts, this);
 
                         StartCoroutine(CallEventCoroutine(derefedCall));
                     }