+++ Runtime/FactExtentions.cs
@@ -16,18 +16,37 @@ namespace ChickadeeEvents
             return null;
         }
 
-        public static void SetValue(this List<Fact> facts, string key, string value)
+        /// <summary>
+        /// Sets the value of the fact in the list with the matching key
+        ///
+        /// If the key doesn't exist yet, it will be created.
+        /// 
+        /// If value is null, the fact will be removed or simply not added
+        /// in the first place.
+        /// </summary>
+        public static void SetValue(this List<Fact> facts, string key,
+                                    string value)
         {
-            foreach (Fact fact in facts)
+            Fact match = facts.Find(e => e.key == key);
+            if(match == null)
             {
-                if (fact.key == key)
-                {
-                    fact.value = value;
-                }
+                if(value != null)
+                    facts.Add(new Fact(key, value));
+                return;
             }
-            facts.Add(new Fact(key, value));
+            if(value == null)
+            {
+                facts.Remove(match);
+                return;
+            }
+            match.value = value;
         }
 
+        /// <summary>
+        /// Gets the value of the fact with the matching key in the list
+        ///
+        /// If the key isn't found, null is returned
+        /// </summary>
         public static Fact GetFact(this List<Fact> facts, string key)
         {
             foreach (Fact fact in facts)