+++ CHANGELOG.md
@@ -5,9 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.1] - 2022-2-11
 
-## [1.1.0] - 2021-2-9
+### Added
+- Expanded debug log section
+- Caller field in EventCall
+
+### Changed
+- Initial event call and rule values
+
+### Fixed
+- List<Fact>.SetValue bug
+- Removed test event call in EventManager
+
+## [1.1.0] - 2022-2-9
 
 ### Added